+++ Changelog.md
@@ -1,3 +1,8 @@
+# 12.0.1-20260323-2
+* Fix bug in MapModule in Classic.
+* Implement objective tracker in Classic.
+* Changed to combined build for Retail and Classic.
+
 # 12.0.1-20260323-1
 * Declutter map, to avoid too close pins.
 * Compatibility with KalielsTracker.